DELIA-31217 : Return all the connected Devices
Reason for change: Return all the connected Devices. In the beginning, we started with Audio In feature and since only one device can be connected for Audio IN and Audio OUT feature, we lived with this.. Now its time to return all the connected devices.
Test Procedure: By the Spark App
Risks: Medium
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Change-Id: I196b134e95e7729929509e1698babe7dc8b98226